Feat/transaction lines

+++ resources/js/Components/templates/InvoicePaymentOptions.vue
@@ -9,15 +9,18 @@ import { formatMoney } from "@/utils";
 
 const { openModal } = usePaymentModal();
 
-interface Props {
-  invoice: Object;
-  accountsEndpoint: string;
-  allowEdit: boolean;
-}
-
-const props = withDefaults(defineProps<Props>(), {
-  accountsEndpoint: "/api/accounts",
-});
+const props = withDefaults(
+  defineProps<{
+    invoice: IInvoice;
+    accountsEndpoint?: string;
+    allowEdit: boolean;
+    externalActions?: Record<string, any>;
+  }>(),
+  {
+    accountsEndpoint: "/api/accounts",
+    allowEdit: false,
+  }
+);
 
 const emit = defineEmits(["edit"]);
 
@@ -108,27 +111,40 @@ const onDelete = async (invoice: IInvoice) => {
 };
 
 const handleActions = (actionName: string, invoice: IInvoice) => {
+  const externalActions = props.externalActions;
   switch (actionName) {
     case "payment":
-      onPayment(invoice);
+      if (externalActions?.payment) {
+        externalActions?.payment?.(invoice);
+      } else {
+        onPayment(invoice);
+      }
       break;
     case "download":
-      onDownload(invoice);
+      if (externalActions?.download) {
+        externalActions?.download?.(invoice);
+      } else {
+        onDownload(invoice);
+      }
       break;
     case "edit":
-      emit("edit");
+      if (externalActions?.edit) {
+        externalActions?.edit?.(invoice);
+      } else {
+        emit("edit");
+      }
       break;
     case "view":
-      router.visit(`/invoices/${invoice.id}`);
+      externalActions?.view?.(invoice) ?? router.visit(`/invoices/${invoice.id}`);
       break;
     case "delete":
-      onDelete(invoice);
+      if (externalActions?.delete) {
+        externalActions?.delete?.(invoice);
+      } else {
+        onDelete(invoice);
+      }
       break;
   }
-};
-
-const refresh = () => {
-  router.reload();
 };
 </script>
 
@@ -140,7 +156,7 @@ const refresh = () => {
     <template #dropdown>
       <ElDropdownMenu>
         <ElDropdownItem :command="actionName" v-for="(action, actionName) in actions">
-          {{ $t(action.label) }}
+          {{ $t(action?.label) }}
         </ElDropdownItem>
       </ElDropdownMenu>
     </template>
